Move screen out of state and into graphics.
Intuitively the screen is a graphics thing. The screen dimensions matter to the game rules so those exist outside, but anything related to the visualization is strictly graphics related. We want t obe able to simply unplug that system and plug in another if we'd like, and that's done by keeping a tight border.